LayoutTests/fast/js/modify-non-references-expected.txt

PASS function f() { g()++; } f.toString() is 'function f() \n{\n  g()++;\n}'
PASS function f() { g()--; } f.toString() is 'function f() \n{\n  g()--;\n}'
PASS function f() { ++g(); } f.toString() is 'function f() \n{\n  ++g();\n}'
PASS function f() { --g(); } f.toString() is 'function f() \n{\n  --g();\n}'
PASS function f() { g() = 1; } f.toString() is 'function f() \n{\n  g() = 1;\n}'
PASS function f() { g() += 1; } f.toString() is 'function f() \n{\n  g() += 1;\n}'
PASS g()++ threw exception ReferenceError: Postfix ++ operator applied to value that is not a reference..
PASS g()-- threw exception ReferenceError: Postfix -- operator applied to value that is not a reference..
PASS ++g() threw exception ReferenceError: Prefix ++ operator applied to value that is not a reference..
PASS --g() threw exception ReferenceError: Prefix -- operator applied to value that is not a reference..
PASS g() = 1 threw exception ReferenceError: Left side of assignment is not a reference..
PASS g() += 1 threw exception ReferenceError: Left side of assignment is not a reference..
PASS successfullyParsed is true

TEST COMPLETE
